+++ web/src/routes/(dashboard)/+layout.server.ts
@@ -1,11 +1,17 @@
 import type { LayoutServerLoad } from './$types';
+import { config } from '$lib/server/config';
 
 export const load: LayoutServerLoad = async ({ locals }) => {
+
+	console.log(config);
+
 	return {
 		user: {
 			name: locals.user?.name,
-			email: locals.user?.email
+			email: locals.user?.email,
+			role: locals.user?.role
 		},
-		locale: locals.locale
+		locale: locals.locale,
+		version: config.version
 	};
 };